// types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IArticleFields {
  /** Title */
  title: string;

  /** content */
  content: Document;

  /** cover image */
  coverImage: Asset;

  /** slug */
  slug: string;

  /** description */
  description: string;

  /** related_articles */
  relatedArticles?: IArticle[] | undefined;

  /** countries */
  countries?: ICountry[] | undefined;

  /** industries */
  industries?: IIndustry[] | undefined;
}

export interface IArticle extends Entry<IArticleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "article";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ICountryFields {
  /** name */
  name: string;

  /** code */
  code: string;
}

export interface ICountry extends Entry<ICountryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "country";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IIndustryFields {
  /** name */
  name: string;
}

export interface IIndustry extends Entry<IIndustryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "industry";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "article" | "country" | "industry";

export type IEntry = IArticle | ICountry | IIndustry;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";